feat(suite): migrate wallet-sdk-solana staking
