+++ suite-common/wallet-types/package.json
@@ -12,6 +12,7 @@
     "dependencies": {
         "@suite-common/intl-types": "workspace:*",
         "@suite-common/metadata-types": "workspace:*",
+        "@suite-common/staking-solana": "workspace:*",
         "@suite-common/suite-config": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
         "@suite-common/wallet-constants": "workspace:*",
