+++ suite-common/wallet-types/tsconfig.json
@@ -4,6 +4,7 @@
     "references": [
         { "path": "../intl-types" },
         { "path": "../metadata-types" },
+        { "path": "../staking-solana" },
         { "path": "../suite-config" },
         { "path": "../wallet-config" },
         { "path": "../wallet-constants" },
