+++ suite-common/wallet-utils/package.json
@@ -12,7 +12,6 @@
         "test-unit:watch": "yarn g:jest -c ../../jest.config.base.js -o --watch"
     },
     "dependencies": {
-        "@everstake/wallet-sdk-solana": "2.1.1",
         "@suite-common/fiat-services": "workspace:*",
         "@suite-common/metadata-types": "workspace:*",
         "@suite-common/suite-config": "workspace:*",
